Move VectorChildBuffer
