+++ src/include/duckdb/common/types/vector.hpp
@@ -233,23 +233,6 @@ protected:
 	mutable buffer_ptr<VectorBuffer> auxiliary;
 };
 
-//! The VectorChildBuffer holds a child Vector
-class VectorChildBuffer : public VectorBuffer {
-public:
-	explicit VectorChildBuffer(Vector vector)
-	    : VectorBuffer(VectorBufferType::VECTOR_CHILD_BUFFER), data(std::move(vector)) {
-	}
-
-public:
-	Vector data;
-	//! Optional size/id to uniquely identify re-occurring dictionaries
-	optional_idx size;
-	string id;
-	//! For caching the hashes of a child buffer
-	mutex cached_hashes_lock;
-	unique_ptr<Vector> cached_hashes;
-};
-
 } // namespace duckdb
 
 #include "duckdb/common/vector/vector_iterator.hpp"
