+++ src/include/duckdb/parser/statement/update_statement.hpp
@@ -51,6 +51,7 @@ public:
 
 public:
 	UpdateStatement();
+	~UpdateStatement() override;
 
 	unique_ptr<UpdateQueryNode> node;
